partnership-recommendation-ai/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.1.3
pandas==2.1.3
numpy==1.23.5
supabase==2.0.3
python-dotenv==1.0.0
websockets==10.4